agentic_ai_backend/agents/conversationagent/pyproject.toml
[project]
name = "conversationagent"
version = "0.1.0"
description = "NR Sector CSS AI Conversation Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework-core>=1.0.0b251211",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.27.1",
    "fastapi>=0.127.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

